demo/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch>=2.0
numpy
scipy
Pillow
h5py
